fix: narrow sqrt() result to bigint for extra-bigint v1.3.0
extra-bigint v1.3.0 changed the return type of sqrt() from bigint to
bigint | null, which broke the dts build with four TS18047 errors in
elegant.ts.

sqrt() only returns null for negative input, and the surrounding guard
already throws for z < 0, so the assertion is safe and runtime behaviour
is unchanged.

+++ lib/elegant.ts
@@ -90,7 +90,8 @@ function decode(z: unknown): unknown {
       throw new TypeError('Invalid argument; negative numbers are not allowed');
     }
 
-    const zSqrt = sqrt(z);
+    // note: sqrt() returns null for negative numbers only, which is ruled out above
+    const zSqrt = sqrt(z) as bigint;
     // note: zSqrt is already floored, bigint does not support floats (dah!)
     const zSqrtSqr = zSqrt * zSqrt;
     const zMinusZSqrtSqr = z - zSqrtSqr;
